Remove sov-stf-runner
