+++ crates/sovereign-sdk/full-node/sov-stf-runner/src/lib.rs
@@ -17,16 +17,17 @@ use std::path::Path;
 #[cfg(feature = "native")]
 use anyhow::Context;
 #[cfg(feature = "native")]
-pub use prover_service::*;
-#[cfg(feature = "native")]
-mod runner;
-#[cfg(feature = "native")]
 pub use config::*;
 #[cfg(feature = "native")]
-pub use runner::*;
+pub use prover_service::*;
+use sov_modules_api::{DaSpec, Zkvm};
+use sov_rollup_interface::stf::StateTransitionFunction;
 
 /// Implements the `StateTransitionVerifier` type for checking the validity of a state transition
 pub mod verifier;
+
+type GenesisParams<ST, Vm, Da> = <ST as StateTransitionFunction<Vm, Da>>::GenesisParams;
+type SoftConfirmationHash = [u8; 32];
 
 #[cfg(feature = "native")]
 /// Reads json file.
@@ -41,4 +42,14 @@ pub fn read_json_file<T: serde::de::DeserializeOwned, P: AsRef<Path>>(
         .with_context(|| format!("Failed to parse genesis from {}", path_str))?;
 
     Ok(config)
+}
+
+#[cfg(feature = "native")]
+/// How [`StateTransitionRunner`] is initialized
+pub enum InitVariant<Stf: StateTransitionFunction<Vm, Da>, Vm: Zkvm, Da: DaSpec> {
+    /// From given state root and soft confirmation hash
+    Initialized((Stf::StateRoot, SoftConfirmationHash)),
+    /// From empty state root
+    /// Genesis params for Stf::init
+    Genesis(GenesisParams<Stf, Vm, Da>),
 }
