+++ crates/sovereign-sdk/full-node/sov-stf-runner/src/prover_service/mod.rs
@@ -5,8 +5,11 @@ use serde::{Deserialize, Serialize};
 use sov_modules_api::Zkvm;
 use sov_rollup_interface::da::DaSpec;
 use sov_rollup_interface::services::da::DaService;
-use sov_rollup_interface::zk::{Proof, StateTransitionData};
+use sov_rollup_interface::stf::StateTransitionFunction;
+use sov_rollup_interface::zk::{Proof, StateTransitionData, ZkvmHost};
 use thiserror::Error;
+
+use crate::verifier::StateTransitionVerifier;
 
 /// The possible configurations of the prover.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -113,4 +116,20 @@ pub trait ProverService<Vm: Zkvm> {
         block_header_hash: <<Self::DaService as DaService>::Spec as DaSpec>::SlotHash,
         da_service: &Self::DaService,
     ) -> Result<(<Self::DaService as DaService>::TransactionId, Proof), anyhow::Error>;
+}
+
+/// Represents the possible modes of execution for a zkVM program
+pub enum ProofGenConfig<Stf, Da: DaService, Vm: ZkvmHost>
+where
+    Stf: StateTransitionFunction<Vm::Guest, Da::Spec>,
+{
+    /// Skips proving.
+    Skip,
+    /// The simulator runs the rollup verifier logic without even emulating the zkVM
+    Simulate(StateTransitionVerifier<Stf, Da::Verifier, Vm::Guest>),
+    /// The executor runs the rollup verification logic in the zkVM, but does not actually
+    /// produce a zk proof
+    Execute,
+    /// The prover runs the rollup verification logic in the zkVM and produces a zk proof
+    Prover,
 }
