+++ crates/sovereign-sdk/module-system/sov-modules-rollup-blueprint/src/lib.rs
@@ -3,23 +3,17 @@
 
 mod runtime_rpc;
 mod wallet;
-use std::net::SocketAddr;
 
 use async_trait::async_trait;
 pub use runtime_rpc::*;
 use sov_db::ledger_db::LedgerDB;
 use sov_modules_api::{Context, DaSpec, Spec};
-use sov_modules_stf_blueprint::{GenesisParams, Runtime as RuntimeTrait, StfBlueprint};
+use sov_modules_stf_blueprint::{GenesisParams, Runtime as RuntimeTrait};
 use sov_rollup_interface::services::da::DaService;
 use sov_rollup_interface::storage::HierarchicalStorageManager;
 use sov_rollup_interface::zk::{Zkvm, ZkvmHost};
-use sov_state::storage::NativeStorage;
 use sov_state::Storage;
-use sov_stf_runner::{
-    FullNodeConfig, InitVariant, ProverConfig, ProverService, StateTransitionRunner,
-};
-use tokio::sync::oneshot;
-use tracing::{instrument, Instrument};
+use sov_stf_runner::{FullNodeConfig, ProverConfig, ProverService};
 pub use wallet::*;
 
 /// This trait defines how to crate all the necessary dependencies required by a rollup.
@@ -123,135 +117,5 @@ pub trait RollupBlueprint: Sized + Send + Sync {
     /// Creates instance of a LedgerDB.
     fn create_ledger_db(&self, rollup_config: &FullNodeConfig<Self::DaConfig>) -> LedgerDB {
         LedgerDB::with_path(&rollup_config.storage.path).expect("Ledger DB failed to open")
-    }
-
-    /// Creates a new rollup.
-    #[instrument(level = "trace", skip_all)]
-    async fn create_new_rollup(
-        &self,
-        runtime_genesis_paths: &<Self::NativeRuntime as RuntimeTrait<
-            Self::NativeContext,
-            Self::DaSpec,
-        >>::GenesisPaths,
-        rollup_config: FullNodeConfig<Self::DaConfig>,
-    ) -> Result<FullNode<Self>, anyhow::Error>
-    where
-        <Self::NativeContext as Spec>::Storage: NativeStorage,
-    {
-        let da_service = self.create_da_service(&rollup_config).await;
-
-        // TODO: Double check what kind of storage needed here.
-        // Maybe whole "prev_root" can be initialized inside runner
-        // Getting block here, so prover_service doesn't have to be `Send`
-
-        let ledger_db = self.create_ledger_db(&rollup_config);
-        let genesis_config = self.create_genesis_config(runtime_genesis_paths, &rollup_config)?;
-
-        let mut storage_manager = self.create_storage_manager(&rollup_config)?;
-        let prover_storage = storage_manager.create_finalized_storage()?;
-
-        let head_soft_batch = ledger_db.get_head_soft_batch()?;
-        let prev_data = match head_soft_batch {
-            Some((number, soft_batch)) => {
-                Some((prover_storage.get_root_hash(number.0 + 1)?, soft_batch.hash))
-            }
-            None => None,
-        };
-
-        let runner_config = rollup_config.runner.expect("Runner config is missing");
-        // TODO(https://github.com/Sovereign-Labs/sovereign-sdk/issues/1218)
-        let rpc_methods = self.create_rpc_methods(
-            &prover_storage,
-            &ledger_db,
-            &da_service,
-            Some(runner_config.sequencer_client_url.clone()),
-        )?;
-
-        let native_stf = StfBlueprint::new();
-
-        let genesis_root = prover_storage.get_root_hash(1);
-
-        let init_variant = match prev_data {
-            Some((root_hash, batch_hash)) => InitVariant::Initialized((root_hash, batch_hash)),
-            None => match genesis_root {
-                Ok(root_hash) => InitVariant::Initialized((root_hash, [0; 32])),
-                _ => InitVariant::Genesis(genesis_config),
-            },
-        };
-
-        let code_commitment = self.get_code_commitment();
-
-        let runner = StateTransitionRunner::new(
-            runner_config,
-            rollup_config.public_keys,
-            rollup_config.rpc,
-            da_service,
-            ledger_db,
-            native_stf,
-            storage_manager,
-            init_variant,
-            None,
-            None,
-            code_commitment,
-        )?;
-
-        Ok(FullNode {
-            runner,
-            rpc_methods,
-        })
-    }
-}
-
-/// Dependencies needed to run the rollup.
-pub struct FullNode<S: RollupBlueprint> {
-    /// The State Transition Runner.
-    #[allow(clippy::type_complexity)]
-    pub runner: StateTransitionRunner<
-        StfBlueprint<S::NativeContext, S::DaSpec, S::Vm, S::NativeRuntime>,
-        S::StorageManager,
-        S::DaService,
-        S::Vm,
-        S::ProverService,
-        S::NativeContext,
-    >,
-    /// Rpc methods for the rollup.
-    pub rpc_methods: jsonrpsee::RpcModule<()>,
-}
-
-impl<S: RollupBlueprint> FullNode<S> {
-    /// Runs the rollup.
-    #[instrument(
-        name = "FullNode",
-        level = "info",
-        skip(self),
-        err,
-        ret(level = "error")
-    )]
-    pub async fn run(self) -> Result<(), anyhow::Error> {
-        self.run_and_report_rpc_port(None).await
-    }
-
-    /// Only run the rpc.
-    pub async fn run_rpc(self) -> Result<(), anyhow::Error> {
-        self.runner
-            .start_rpc_server(self.rpc_methods, None)
-            .instrument(tracing::Span::current())
-            .await;
-        Ok(())
-    }
-
-    /// Runs the rollup. Reports rpc port to the caller using the provided channel.
-    pub async fn run_and_report_rpc_port(
-        self,
-        channel: Option<oneshot::Sender<SocketAddr>>,
-    ) -> Result<(), anyhow::Error> {
-        let mut runner = self.runner;
-        runner
-            .start_rpc_server(self.rpc_methods, channel)
-            .instrument(tracing::Span::current())
-            .await;
-
-        runner.run_in_process().await?;
-        Ok(())
     }
 }
